forc init --template removed, pkg.rs unused pub modifier removed
