+++ forc/src/ops/forc_template.rs
@@ -1,17 +1,20 @@
 use crate::cli::TemplateCommand;
-use crate::ops::forc_init::{edit_cargo_toml, edit_forc_toml};
 use crate::utils::{defaults, SWAY_GIT_TAG};
-use anyhow::{anyhow, Result};
+use anyhow::{anyhow, Context, Result};
 use forc_pkg::{
     fetch_git, fetch_id, find_dir_within, git_commit_path, pin_git, Manifest, SourceGit,
 };
+use forc_util::validate_name;
 use fs_extra::dir::{copy, CopyOptions};
-use std::path::PathBuf;
+use std::fs::File;
+use std::io::{Read, Write};
+use std::path::{Path, PathBuf};
 use std::{env, fs};
 use sway_utils::constants;
 use url::Url;
 
 pub fn init(command: TemplateCommand) -> Result<()> {
+    validate_name(&command.project_name, "project name")?;
     let template_name = match command.template_name.clone() {
         Some(temp_name) => temp_name,
         None => "DEFAULT_TEMP_NAME".to_string(),
@@ -81,6 +84,79 @@ pub fn init(command: TemplateCommand) -> Result<()> {
             defaults::default_tests_manifest(&command.project_name),
         )?;
     }
+    Ok(())
+}
+
+fn edit_forc_toml(out_dir: &Path, project_name: &str, real_name: &str) -> Result<()> {
+    let mut file = File::open(out_dir.join(constants::MANIFEST_FILE_NAME))?;
+    let mut toml = String::new();
+    file.read_to_string(&mut toml)?;
+    let mut manifest_toml = toml.parse::<toml_edit::Document>()?;
+
+    let mut authors = Vec::new();
+    let forc_toml: toml::Value = toml::de::from_str(&toml)?;
+    if let Some(table) = forc_toml.as_table() {
+        if let Some(package) = table.get("project") {
+            // If authors Vec is currently populated use that
+            if let Some(toml::Value::Array(authors_vec)) = package.get("authors") {
+                for author in authors_vec {
+                    if let toml::value::Value::String(name) = &author {
+                        authors.push(name.clone());
+                    }
+                }
+            }
+        }
+    }
+
+    // Only append the users name to the authors field if it isn't already in the list
+    if authors.iter().any(|e| e != real_name) {
+        authors.push(real_name.to_string());
+    }
+
+    let authors: toml_edit::Array = authors.iter().collect();
+    manifest_toml["project"]["authors"] = toml_edit::value(authors);
+    manifest_toml["project"]["name"] = toml_edit::value(project_name);
+
+    // Remove explicit std entry from copied template
+    if let Some(project) = manifest_toml.get_mut("dependencies") {
+        let _ = project
+            .as_table_mut()
+            .context("Unable to get forc manifest as table")?
+            .remove("std");
+    }
+
+    let mut file = File::create(out_dir.join(constants::MANIFEST_FILE_NAME))?;
+    file.write_all(manifest_toml.to_string().as_bytes())?;
+    Ok(())
+}
+
+fn edit_cargo_toml(out_dir: &Path, project_name: &str, real_name: &str) -> Result<()> {
+    let mut file = File::open(out_dir.join(constants::TEST_MANIFEST_FILE_NAME))?;
+    let mut toml = String::new();
+    file.read_to_string(&mut toml)?;
+
+    let mut updated_authors = toml_edit::Array::default();
+
+    let cargo_toml: toml::Value = toml::de::from_str(&toml)?;
+    if let Some(table) = cargo_toml.as_table() {
+        if let Some(package) = table.get("package") {
+            if let Some(toml::Value::Array(authors_vec)) = package.get("authors") {
+                for author in authors_vec {
+                    if let toml::value::Value::String(name) = &author {
+                        updated_authors.push(name);
+                    }
+                }
+            }
+        }
+    }
+    updated_authors.push(real_name);
+
+    let mut manifest_toml = toml.parse::<toml_edit::Document>()?;
+    manifest_toml["package"]["authors"] = toml_edit::value(updated_authors);
+    manifest_toml["package"]["name"] = toml_edit::value(project_name);
+
+    let mut file = File::create(out_dir.join(constants::TEST_MANIFEST_FILE_NAME))?;
+    file.write_all(manifest_toml.to_string().as_bytes())?;
     Ok(())
 }
